requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.7.4
Django==5.0.3
django-debug-toolbar==4.4.6
django-filter==24.2
djangorestframework==3.15.2
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.1.0
flake8-docstrings==1.7.0
future==0.18.3
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pefile==2021.9.3
pewee==0.1.0.dev9
pillow==10.3.0
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sentry-sdk==2.10.0
snowballstemmer==2.2.0
sqlparse==0.4.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.2
